backend/requirements.txt
asgiref==3.8.1
Django==4.2.15
django-cors-headers==4.4.0
djangorestframework==3.15.2
sqlparse==0.5.1
psycopg2==2.9.9
requests==2.32.3
drf-yasg==1.21.7